morello: Fix kselftest build
A recent patch introduced the possibility to build kselftest separately
from the kernel. This patch introduced a preliminary step useful to
configure the kernel repository.

If the kselftest are built separately from the kernel the configuration
step does not happen resulting in a failure.

Fix the kselftest build making sure that the configuration script is
invoked before the compilation in the case in which it was not
previously called.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>