Fix issues with scripts, install to a non-root location
Scripts to build libraries simply didn't work, because they did not have
a correct PATH. Another bug in mistaking OPTIONS_BUILD_LIB and BUILD_LIB
also prevented them from being built as intended.

The other change made was to remove the restriction that the --install
option can only be used in a container by installing to a user's home
directory instead of the root directory.